+++ feature/calendar/src/main/java/com/ku_stacks/ku_ring/calendar/compose/component/calendar/CalendarDefaults.kt
@@ -1,4 +1,4 @@
-package com.ku_stacks.ku_ring.calendar.model
+package com.ku_stacks.ku_ring.calendar.compose.component.calendar
 
 import androidx.compose.runtime.Composable
 import com.ku_stacks.ku_ring.designsystem.kuringtheme.KuringTheme
